Additions to text over paths
Added a utility routine "body_path_calibrator()" that can be used to
determine the parameters for the body.rising_text, body.setting_text
etc.

Added an option to bodies to determine text size. So far, there are only
two sizes but it is extensible

Added paths for two deep sky objects (M45 and Antares), demonstrating
smaller sizes.

A few arrangements for the Solar System objects to remove conflicts.